Add some documentation to load asset env var with right data dir
It would be nearly impossible to hit the execution root, but what we
can do is add to the data dir and then reference via runfiles. Despite
the documentation I haven't found that we get a relative path to the
target either for the execution root or the data dir, so this is the
best we've found.

https://bazel.build/extending/rules#runfiles_location
https://github.com/bazelbuild/bazel/issues/15985